UBLE-151 refactor: 파일명 오탈자 수정
